+++ OrchestratorV2/interfaces/common/releasefilter.ts
@@ -1,7 +1,10 @@
+import { EnvironmentStatus } from "azure-devops-node-api/interfaces/ReleaseInterfaces";
+
 export interface IReleaseFilter {
 
     artifactVersion?: string;
     sourceBranch?: string;
     tag?: string[];
+    stageStatus?: EnvironmentStatus[],
 
 }