+++ OrchestratorV2/interfaces/task/filters.ts
@@ -3,5 +3,6 @@ export interface IFilters {
     releaseTags: string[];
     artifactTags: string[];
     artifactBranch?: string;
+    stageStatuses: string[];
 
 }